tests/wpt/mozilla/tests/mozilla/worklets/test-paint-worklet-loading.html

<!doctype html>
<meta charset="utf-8">
<title>Test paint worklet loading</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
var host_info = get_host_info();

promise_test(function() {
  return CSS.paintWorklet.addModule("test-paint-worklet.js");
}, "Loading a paint worklet.");

promise_test(function(t) {
   var path = new URL("test_paint_worklet.js", document.location).pathname;
   var url = new URL(path, host_info.HTTP_REMOTE_ORIGIN);
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule(url));
}, "Loading a cross-origin paint worklet.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("nonexistent-worklet.js"));
}, "Loading a nonexistent paint worklet.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("syntax-error.js"));
}, "Loading a syntactically incorrect paint worklet.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("throw-exception.js"));
}, "Loading an exception-throwing paint worklet.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet.js"));
}, "Loading a paint worklet again.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-empty-name.js"));
}, "Loading a paint worklet with an empty name.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-properties-throws.js"));
}, "Loading a paint worklet whose inputProperties throws an exception.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-arguments-throws.js"));
}, "Loading a paint worklet whose inputArguments throws an exception.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-alpha-throws.js"));
}, "Loading a paint worklet whose alpha throws an exception.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-not-constructor.js"));
}, "Loading a paint worklet which isn't a constructor function.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-no-paint.js"));
}, "Loading a paint worklet with no paint.");

promise_test(function(t) {
   return promise_rejects_dom(t, "AbortError", CSS.paintWorklet.addModule("test-paint-worklet-paint-not-callable.js"));
}, "Loading a paint worklet with a paint that is not callable.");
</script>
